% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{format_gwas}
\alias{format_gwas}
\title{Convert GWAS summary statistics into required format}
\usage{
format_gwas(
  gwas,
  chr = "CHR",
  snp = "SNP",
  effect_allele = "A1",
  other_allele = "A2",
  eaf = "eaf",
  beta = "beta",
  n_case = "N_case",
  n_control = "N_cont"
)
}
\arguments{
\item{gwas}{GWAS summary statistics}

\item{chr}{Column name for chromosome - a string}

\item{snp}{Column name for rsid of SNP - a string}

\item{effect_allele}{Column name for effect allele (i.e. the allele for the signed statistic beta) - a string}

\item{other_allele}{Column name for the non-effect allele - a string}

\item{eaf}{Column name for effect allele frequency - a string}

\item{beta}{Column name for per-allele log odds ratio - a string}

\item{n_case}{Column name for number of cases - a string}

\item{n_control}{Column name for number of controls - a string}
}
\value{
a data frame with formatted GWAS summary statistics
}
\description{
Convert GWAS summary statistics into required format
}
\examples{
df = format_gwas(gwas = gwas_sumstats,chr = "CHR",effect_allele = "A1",other_allele = "A2",snp = "SNP",eaf = "eaf",beta = "beta",n_case = "n_case",n_control = "n_control")
df = format_gwas(gwas = gwas_sumstats)
}
