% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gwas_power}
\alias{gwas_power}
\title{Calculate power for GWAS of a binary trait}
\usage{
gwas_power(
  eaf = 0.5,
  beta_alt = 0.5,
  n_case = 1000,
  n_control = 1000,
  alpha = 5e-08,
  n_perm = 1000
)
}
\arguments{
\item{eaf}{effect allele frequency}

\item{beta_alt}{beta under the alternate hypothesis (i.e. per-allele log odds ratio in your discovery GWAS)}

\item{n_case}{number of cases}

\item{n_control}{number of controls}

\item{alpha}{desired alpha}

\item{n_perm}{number of permutations for power calculations}
}
\value{
the estimated power for each SNP given the allele frequency, n, alpha, effect size, and number of permutations. Note that this is for a one-tailed test.
}
\description{
Calculate power for GWAS of a binary trait
}
