% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{power_calc_per_pop}
\alias{power_calc_per_pop}
\title{Power calculation}
\usage{
power_calc_per_pop(gwas, n_perm, alpha, pop)
}
\arguments{
\item{gwas}{GWAS summary statistics with 1kg allele frequencies (output of combine_gwas_with_1kg_freqs)}

\item{n_perm}{Number of permutations, defaults to 100.}

\item{alpha}{Alpha, defaults to 0.05 (replication)}

\item{pop}{A string which must be one of "AFR","SAS","AMR","EAS",or "EUR" specifying the 1kg superpopulation of interest.}
}
\value{
a data frame with your input summary statistics plus a new column listing the power at each SNP in the superpopulation of interest at the desired alpha.
}
\description{
Power calculation
}
\examples{

df = format_gwas(gwas = gwas_sumstats)
kg_combo = combine_gwas_with_1kg_freqs(df)
power_stats = power_calc_per_pop(kg_combo,n_perm = 1000,alpha=0.05,pop="AFR") # for AFR
power_stats = power_calc_per_pop(kg_combo,n_perm = 1000,alpha=0.05,pop="SAS") # for SAS
}
